Different constructors for the simulation class depending on ROF table use (or lack thereof). Simulation exception handling is now in its own function. Default Borg seed was changed to no seed.

+++ src/ContinuityModels/Base/ContinuityModel.h
@@ -38,7 +38,7 @@ protected:
     int delta_realization_weeks[NUMBER_REALIZATIONS_ROF];
 
 public:
-    const unsigned int realization_id;
+    const unsigned long realization_id;
 
     ContinuityModel(vector<WaterSource *> &water_sources, vector<Utility *> &utilities,
                     vector<MinEnvFlowControl *> &min_env_flow_controls,
@@ -46,7 +46,7 @@ public:
                     const vector<vector<int>> &water_sources_to_utilities,
                     vector<double> &utilities_rdm,
                     vector<double> &water_sources_rdm,
-                    unsigned int realization_id);
+                    unsigned long realization_id);
 
     ContinuityModel(ContinuityModel &continuity_model);
 
@@ -56,7 +56,7 @@ public:
 
     virtual ~ContinuityModel();
 
-    void setRealization(unsigned int realization_id, vector<double> &utilities_rdm,
+    void setRealization(unsigned long realization_id, vector<double> &utilities_rdm,
                         vector<double> &water_sources_rdm);
 
     vector<int> getOnlineDownstreamSources();
