+++ src/Controls/Custom/JordanLakeMinEnvFlowControl.cpp
@@ -80,6 +80,6 @@ double JordanLakeMinEnvFlowControl::getRelease(int week) {
         return rel06_up;
 }
 
-void JordanLakeMinEnvFlowControl::setRealization(unsigned int r, vector<double> &rdm_factors) {
+void JordanLakeMinEnvFlowControl::setRealization(unsigned long r, vector<double> &rdm_factors) {
     lillington_gage_catchment.setRealization(r, rdm_factors);
 }
